Fix debugger test
D78585722  broke test_debug and somehow the tests don't run on diffs.

This adds the necessary annotations to make it ok for the  debugger code to block the existing event loop while it is doing debugging stuff.

Differential Revision: [D78634155](https://our.internmc.facebook.com/intern/diff/D78634155/)

**NOTE FOR REVIEWERS**: This PR has internal Meta-specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D78634155/)!

[ghstack-poisoned]
